requirements.txt
streamlit==1.24.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
ultralytics==8.0.124
pandas==2.0.3
python-dotenv==1.0.0